A few updates for server managers.

+++ grow/server/manager.py
@@ -10,7 +10,7 @@ import os
 import yaml
 
 _servers = {}
-_config_path = '{}/.grow/servers'.format(os.environ['HOME'])
+_config_path = '{}/.grow/servers.yaml'.format(os.environ['HOME'])
 
 
 def start(root, port=None, use_subprocess=False):
@@ -22,18 +22,20 @@ def start(root, port=None, use_subprocess=False):
   if not use_subprocess:
     httpserver.serve(main_lib.services_app)
     return
-  logging.info('Started server for pod: {}'.format(root))
   process = multiprocessing.Process(
       target=httpserver.serve,
-      args=(main_lib.services_app,))
+      args=(main_lib.services_app,),
+      kwargs={'port': port})
   process.start()
   _servers[root] = process
+  return process
 
 
 def stop(root):
-  process = _servers[root]
+  process = _servers.pop(root, None)
+  if not process:
+    return
   process.terminate()
-  del _servers[root]
   logging.info('Stopped server for pod: {}'.format(root))
 
 
@@ -41,10 +43,6 @@ def stop(root):
 def stop_all():
   for root in _servers.keys():
     stop(root)
-
-
-def read_config():
-  return yaml.load(_config_path)
 
 
 def write_config(config):
@@ -57,28 +55,40 @@ def write_config(config):
   fp.close()
 
 
-class Pod(object):
 
-  def __init__(self, root, port=8000):
+class PodServer(object):
+
+  def __init__(self, root, port=8000, revision_status=None):
     self.root = root
     self.port = port
+    self.revision_status = revision_status
+    self.server_status = 'off'
+    self._process = None
 
-  @property
-  def revision_status(self):
-    pass
+  def start(self):
+    self._process = start(self.root, port=self.port, use_subprocess=True)
+    self.server_status = 'on'
+    logging.info('Started server for pod: {}'.format(self.root))
 
-  @property
-  def server_status(self):
-    return 'off'
-
-  def start_server(self):
-    pass
-
-  def stop_server(self):
-    pass
+  def stop(self):
+    self.server_status = 'off'
+    self._process.terminate()
+    logging.info('Stopped server for pod: {}'.format(self.root))
 
   def set_root(self, root):
     self.root = root
 
   def set_port(self, port):
     self.port = port
+
+  @classmethod
+  def load(cls):
+    servers = []
+    fp = open(_config_path)
+    for server in yaml.load(fp)['servers']:
+      servers.append(cls(
+          server['root'],
+          port=server.get('port'),
+          revision_status=server.get('revision_status')
+      ))
+    return servers
